# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Trade Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStatus {
  running: boolean;
  /**
     * Seconds since bot started
     * @nullable
     */
  uptime: number | null;
  /** @nullable */
  startedAt: string | null;
}

export interface BotConfig {
  hasXApiKey: boolean;
  hasXApiSecret: boolean;
  hasXAccessToken: boolean;
  hasXAccessSecret: boolean;
  hasOpenAiKey: boolean;
  hasXaiKey: boolean;
  isConfigured: boolean;
}

export interface BotConfigInput {
  xApiKey?: string;
  xApiSecret?: string;
  xAccessToken?: string;
  xAccessSecret?: string;
  openAiKey?: string;
  xaiKey?: string;
}

export type ActivityLogType = typeof ActivityLogType[keyof typeof ActivityLogType];


export const ActivityLogType = {
  post: 'post',
  reply: 'reply',
  influencer_reply: 'influencer_reply',
} as const;

export interface ActivityLog {
  id: number;
  type: ActivityLogType;
  content: string;
  /** @nullable */
  targetUser?: string | null;
  /** @nullable */
  tweetId?: string | null;
  success?: boolean;
  createdAt: string;
}

export interface BotStats {
  totalPosts: number;
  totalReplies: number;
  totalInfluencerReplies: number;
  last24hPosts: number;
  last24hReplies: number;
}

export interface MemoryEntry {
  id: number;
  user: string;
  question: string;
  answer: string;
}

export type GetActivityParams = {
limit?: number;
type?: GetActivityType;
};

export type GetActivityType = typeof GetActivityType[keyof typeof GetActivityType];


export const GetActivityType = {
  post: 'post',
  reply: 'reply',
  influencer_reply: 'influencer_reply',
} as const;