# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexus Trade Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/config:
    get:
      operationId: getBotConfig
      tags: [bot]
      summary: Get bot configuration (keys are masked)
      responses:
        "200":
          description: Bot config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
    put:
      operationId: saveBotConfig
      tags: [bot]
      summary: Save bot configuration and API keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotConfigInput"
      responses:
        "200":
          description: Config saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
  /bot/activity:
    get:
      operationId: getActivity
      tags: [activity]
      summary: Get recent bot activity (posts and replies)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: type
          in: query
          schema:
            type: string
            enum: [post, reply, influencer_reply]
      responses:
        "200":
          description: Activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"
...
components:
  schemas:
...
    BotConfig:
      type: object
      required: [hasXApiKey, hasXApiSecret, hasXAccessToken, hasXAccessSecret, hasOpenAiKey, hasXaiKey, isConfigured]
      properties:
        hasXApiKey:
          type: boolean
        hasXApiSecret:
          type: boolean
        hasXAccessToken:
          type: boolean
        hasXAccessSecret:
          type: boolean
        hasOpenAiKey:
          type: boolean
        hasXaiKey:
          type: boolean
        isConfigured:
          type: boolean
    BotConfigInput:
      type: object
      properties:
        xApiKey:
          type: string
        xApiSecret:
          type: string
        xAccessToken:
          type: string
        xAccessSecret:
          type: string
        openAiKey:
          type: string
        xaiKey:
          type: string
    ActivityLog:
      type: object
      required: [id, type, content, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [post, reply, influencer_reply]
        content:
          type: string
        targetUser:
          type: ["string", "null"]
        tweetId:
          type: ["string", "null"]
        success:
          type: boolean
        createdAt:
          type: string
          format: date-time
...