# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Trade Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityLogType } from './activityLogType';

export interface ActivityLog {
  id: number;
  type: ActivityLogType;
  content: string;
  /** @nullable */
  targetUser?: string | null;
  /** @nullable */
  tweetId?: string | null;
  success?: boolean;
  createdAt: Date;
}