# lib/api-zod/src/generated/types/botConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Trade Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfig {
  hasXApiKey: boolean;
  hasXApiSecret: boolean;
  hasXAccessToken: boolean;
  hasXAccessSecret: boolean;
  hasOpenAiKey: boolean;
  hasXaiKey: boolean;
  isConfigured: boolean;
}