# lib/api-zod/src/generated/types/botConfigInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Trade Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfigInput {
  xApiKey?: string;
  xApiSecret?: string;
  xAccessToken?: string;
  xAccessSecret?: string;
  openAiKey?: string;
  xaiKey?: string;
}